+++ src/node.hh
@@ -14,18 +14,18 @@ class Node
 
     int dist2(Uint8 r, Uint8 g, Uint8 b);
 
-    int getR();
-    int getG();
-    int getB();
+    Uint8 getR();
+    Uint8 getG();
+    Uint8 getB();
 
-    void setR(int r);
-    void setG(int g);
-    void setB(int b);
+    void setR(Uint8 r);
+    void setG(Uint8 g);
+    void setB(Uint8 b);
 
   private:
-    int R;
-    int G;
-    int B;
+    Uint8 R;
+    Uint8 G;
+    Uint8 B;
 };
 
 #endif
